refactor(ical): move getDateKey to internal lib/date-utils module
